refactor(common): CustomException, ErrorInfo, handler 생성
- 예외 처리를 Custom 하게 하기 위해서 CustomException 생성
- 에러 정보를 모으는 ErrorInfo enum 생성
- CustomException handler 작성
