+++ src/main/java/kitchenpos/common/error/CustomException.java
@@ -1,0 +1,7 @@
+package kitchenpos.common.error;
+
+public class CustomException extends RuntimeException{
+    public CustomException(ErrorInfo errorInfo) {
+        super(errorInfo.message());
+    }
+}
