+++ src/main/java/kitchenpos/handler/CustomExceptionHandler.java
@@ -1,0 +1,16 @@
+package kitchenpos.handler;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import kitchenpos.common.error.CustomException;
+
+@RestControllerAdvice
+public class CustomExceptionHandler {
+    @ExceptionHandler(CustomException.class)
+    public ResponseEntity customException() {
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+    }
+}
